Create a model stub
