+++ src/test/java/seedu/smartnus/logic/commands/AddMcqCommandTest.java
@@ -6,31 +6,21 @@ import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static seedu.smartnus.testutil.Assert.assertThrows;
 
-import java.nio.file.Path;
 import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.Comparator;
-import java.util.function.Predicate;
 
 import org.junit.jupiter.api.Test;
 
-import javafx.collections.ObservableList;
-import seedu.smartnus.commons.core.GuiSettings;
 import seedu.smartnus.commons.core.Messages;
-import seedu.smartnus.commons.core.theme.Theme;
 import seedu.smartnus.logic.commands.exceptions.CommandException;
 import seedu.smartnus.logic.commands.questions.AddMcqCommand;
 import seedu.smartnus.logic.commands.questions.AddQuestionCommand;
-import seedu.smartnus.model.Model;
+import seedu.smartnus.model.ModelStub;
+import seedu.smartnus.model.ModelStubTagPanel;
 import seedu.smartnus.model.ReadOnlySmartNus;
-import seedu.smartnus.model.ReadOnlyUserPrefs;
 import seedu.smartnus.model.SmartNus;
-import seedu.smartnus.model.note.Note;
 import seedu.smartnus.model.question.Question;
-import seedu.smartnus.model.statistic.TagStatistic;
 import seedu.smartnus.testutil.QuestionBuilder;
-import seedu.smartnus.ui.panel.NoteListPanel;
-import seedu.smartnus.ui.panel.QuestionListPanel;
 
 class AddMcqCommandTest {
     @Test
@@ -54,7 +44,7 @@ class AddMcqCommandTest {
     public void execute_duplicateQuestion_throwsCommandException() {
         Question validQuestion = new QuestionBuilder().build();
         AddMcqCommand addCommand = new AddMcqCommand(validQuestion);
-        AddMcqCommandTest.ModelStub modelStub = new AddMcqCommandTest.ModelStubWithQuestion(validQuestion);
+        ModelStub modelStub = new AddMcqCommandTest.ModelStubWithQuestion(validQuestion);
 
         assertThrows(CommandException.class,
                 AddQuestionCommand.MESSAGE_DUPLICATE_QUESTION, () -> addCommand.execute(modelStub));
@@ -63,7 +53,7 @@ class AddMcqCommandTest {
     /**
      * A Model stub that contains a single question.
      */
-    private class ModelStubWithQuestion extends AddMcqCommandTest.ModelStub {
+    private class ModelStubWithQuestion extends ModelStub {
         private final Question question;
 
         ModelStubWithQuestion(Question question) {
@@ -104,162 +94,12 @@ class AddMcqCommandTest {
 
     @Test
     public void execute_inWrongPanel() throws Exception {
-        AddMcqCommandTest.ModelStubAcceptingWrongPanel modelStub = new AddMcqCommandTest.ModelStubAcceptingWrongPanel();
+        ModelStubTagPanel modelStub = new ModelStubTagPanel();
         Question validQuestion = new QuestionBuilder().build();
         AddMcqCommand addCommand = new AddMcqCommand(validQuestion);
 
         assertThrows(CommandException.class,
                 Messages.MESSAGE_NOT_IN_QUESTION_PANEL, () -> addCommand.execute(modelStub));
-    }
-
-    /**
-     * A default model stub that have all of the methods failing.
-     */
-    private class ModelStub implements Model {
-        @Override
-        public void setUserPrefs(ReadOnlyUserPrefs userPrefs) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public ReadOnlyUserPrefs getUserPrefs() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public GuiSettings getGuiSettings() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setGuiSettings(GuiSettings guiSettings) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public Path getSmartNusFilePath() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setSmartNusFilePath(Path smartNusFilePath) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void addQuestion(Question question) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setSmartNus(ReadOnlySmartNus smartNus) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public ReadOnlySmartNus getSmartNus() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public boolean hasQuestion(Question question) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void deleteQuestion(Question target) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setQuestion(Question target, Question editedQuestion) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public ObservableList<Question> getFilteredQuestionList() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public ObservableList<Question> getFilteredQuizQuestionList() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void updateFilteredQuestionList(Predicate<Question> predicate) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void resetFilteredQuestionList() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void sortFilteredQuizQuestionList(Comparator<Question> comparator) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setTheme(Theme theme) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public Theme getTheme() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public ObservableList<TagStatistic> getTagStatistic() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void updateFilteredTagStatistic(Predicate<TagStatistic> predicate) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setPanel(String panel) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public String getPanel() {
-            return QuestionListPanel.QUESTION_PANEL;
-        }
-
-        @Override
-        public void addNote(Note note) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void deleteNote(Note target) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setNote(Note target, Note editedNote) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public ObservableList<Note> getFilteredNoteList() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void updateFilteredNoteList(Predicate<Note> predicate) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void updateFilteredQuizQuestionList(Predicate<Question> predicate) {
-            throw new AssertionError("This method should not be called.");
-        }
     }
 
     /**
@@ -283,17 +123,6 @@ class AddMcqCommandTest {
         @Override
         public ReadOnlySmartNus getSmartNus() {
             return new SmartNus();
-        }
-    }
-
-    /**
-     * A Model stub that is not in the question panel.
-     */
-    private class ModelStubAcceptingWrongPanel extends AddMcqCommandTest.ModelStub {
-
-        @Override
-        public String getPanel() {
-            return NoteListPanel.NOTE_PANEL;
         }
     }
 }
