+++ src/test/java/seedu/smartnus/logic/commands/AddNoteCommandTest.java
@@ -4,29 +4,20 @@ import static java.util.Objects.requireNonNull;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static seedu.smartnus.testutil.Assert.assertThrows;
 
-import java.nio.file.Path;
 import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.Comparator;
-import java.util.function.Predicate;
 
 import org.junit.jupiter.api.Test;
 
-import javafx.collections.ObservableList;
-import seedu.smartnus.commons.core.GuiSettings;
 import seedu.smartnus.commons.core.Messages;
-import seedu.smartnus.commons.core.theme.Theme;
 import seedu.smartnus.logic.commands.exceptions.CommandException;
-import seedu.smartnus.model.Model;
+import seedu.smartnus.model.ModelStub;
+import seedu.smartnus.model.ModelStubTagPanel;
 import seedu.smartnus.model.ReadOnlySmartNus;
-import seedu.smartnus.model.ReadOnlyUserPrefs;
 import seedu.smartnus.model.SmartNus;
 import seedu.smartnus.model.note.Note;
-import seedu.smartnus.model.question.Question;
-import seedu.smartnus.model.statistic.TagStatistic;
 import seedu.smartnus.testutil.NoteBuilder;
 import seedu.smartnus.ui.panel.NoteListPanel;
-import seedu.smartnus.ui.panel.QuestionListPanel;
 
 public class AddNoteCommandTest {
     @Test
@@ -47,7 +38,7 @@ public class AddNoteCommandTest {
 
     @Test
     public void execute_inWrongPanel() throws Exception {
-        ModelStubAcceptingWrongPanel modelStub = new ModelStubAcceptingWrongPanel();
+        ModelStubTagPanel modelStub = new ModelStubTagPanel();
         Note validNote = new NoteBuilder().build();
 
 
@@ -57,158 +48,16 @@ public class AddNoteCommandTest {
 
     // todo: if duplicate note checker is implemented, add a test here to check if it works properly
 
-    private class ModelStub implements Model {
-        @Override
-        public void setUserPrefs(ReadOnlyUserPrefs userPrefs) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public ReadOnlyUserPrefs getUserPrefs() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public GuiSettings getGuiSettings() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setGuiSettings(GuiSettings guiSettings) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public Path getSmartNusFilePath() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setSmartNusFilePath(Path smartNusFilePath) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void addQuestion(Question question) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setSmartNus(ReadOnlySmartNus smartNus) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public ReadOnlySmartNus getSmartNus() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public boolean hasQuestion(Question question) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void deleteQuestion(Question target) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setQuestion(Question target, Question editedQuestion) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public ObservableList<Question> getFilteredQuestionList() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public ObservableList<Question> getFilteredQuizQuestionList() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void updateFilteredQuestionList(Predicate<Question> predicate) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void resetFilteredQuestionList() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void addNote(Note note) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void deleteNote(Note target) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setNote(Note target, Note editedNote) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public ObservableList<Note> getFilteredNoteList() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void updateFilteredNoteList(Predicate<Note> predicate) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void updateFilteredQuizQuestionList(Predicate<Question> predicate) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void sortFilteredQuizQuestionList(Comparator<Question> comparator) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setTheme(Theme theme) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public Theme getTheme() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setPanel(String panel) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public String getPanel() {
-            return NoteListPanel.NOTE_PANEL;
-        }
-
-        @Override
-        public ObservableList<TagStatistic> getTagStatistic() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void updateFilteredTagStatistic(Predicate<TagStatistic> predicate) {
-            throw new AssertionError("This method should not be called.");
-        }
-    }
-
     /**
      * A Model stub that always accept the question being added.
      */
     private class ModelStubAcceptingNoteAdded extends ModelStub {
         final ArrayList<Note> notesAdded = new ArrayList<>();
+
+        @Override
+        public String getPanel() {
+            return NoteListPanel.NOTE_PANEL;
+        }
 
         @Override
         public void addNote(Note note) {
@@ -219,18 +68,6 @@ public class AddNoteCommandTest {
         @Override
         public ReadOnlySmartNus getSmartNus() {
             return new SmartNus();
-        }
-    }
-
-    /**
-     * A Model stub that is not in the notes panel.
-     */
-    private class ModelStubAcceptingWrongPanel extends ModelStub {
-        final ArrayList<Note> notesAdded = new ArrayList<>();
-
-        @Override
-        public String getPanel() {
-            return QuestionListPanel.QUESTION_PANEL;
         }
     }
 }
