+++ src/test/java/seedu/smartnus/logic/commands/EditCommandTest.java
@@ -18,39 +18,29 @@ import static seedu.smartnus.testutil.TypicalIndexes.INDEX_FIRST_QUESTION;
 import static seedu.smartnus.testutil.TypicalIndexes.INDEX_SECOND_QUESTION;
 import static seedu.smartnus.testutil.TypicalSmartNus.getTypicalSmartNus;
 
-import java.nio.file.Path;
-import java.util.Comparator;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
-import java.util.function.Predicate;
 
 import org.junit.jupiter.api.Test;
 
-import javafx.collections.ObservableList;
-import seedu.smartnus.commons.core.GuiSettings;
 import seedu.smartnus.commons.core.Messages;
 import seedu.smartnus.commons.core.index.Index;
-import seedu.smartnus.commons.core.theme.Theme;
 import seedu.smartnus.logic.commands.EditCommand.EditQuestionDescriptor;
 import seedu.smartnus.logic.commands.exceptions.CommandException;
 import seedu.smartnus.model.Model;
 import seedu.smartnus.model.ModelManager;
-import seedu.smartnus.model.ReadOnlySmartNus;
-import seedu.smartnus.model.ReadOnlyUserPrefs;
+import seedu.smartnus.model.ModelStubTagPanel;
 import seedu.smartnus.model.SmartNus;
 import seedu.smartnus.model.UserPrefs;
 import seedu.smartnus.model.choice.Choice;
-import seedu.smartnus.model.note.Note;
 import seedu.smartnus.model.question.MultipleChoiceQuestion;
 import seedu.smartnus.model.question.Question;
 import seedu.smartnus.model.question.ShortAnswerQuestion;
 import seedu.smartnus.model.question.TrueFalseQuestion;
-import seedu.smartnus.model.statistic.TagStatistic;
 import seedu.smartnus.testutil.EditQuestionDescriptorBuilder;
 import seedu.smartnus.testutil.QuestionBuilder;
 import seedu.smartnus.testutil.TypicalQuestions;
-import seedu.smartnus.ui.panel.StatisticListPanel;
 
 /**
  * Contains integration tests (interaction with the Model) and unit tests for EditCommand.
@@ -292,160 +282,10 @@ public class EditCommandTest {
 
     @Test
     public void edit_wrongPanel() {
-        ModelStub model = new EditCommandTest.ModelStub();
+        ModelStubTagPanel model = new ModelStubTagPanel();
         EditCommand qnCmd = new EditCommand(INDEX_FIRST_QUESTION, DESC_AMY);
         assertThrows(CommandException.class,
                 Messages.MESSAGE_NOT_IN_QUESTION_PANEL, () -> qnCmd.execute(model));
-    }
-
-    /**
-     * A default model stub that have all of the methods failing.
-     */
-    private class ModelStub implements Model {
-        @Override
-        public void setUserPrefs(ReadOnlyUserPrefs userPrefs) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public ReadOnlyUserPrefs getUserPrefs() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public GuiSettings getGuiSettings() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setGuiSettings(GuiSettings guiSettings) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public Path getSmartNusFilePath() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setSmartNusFilePath(Path smartNusFilePath) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void addQuestion(Question question) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setSmartNus(ReadOnlySmartNus smartNus) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public ReadOnlySmartNus getSmartNus() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public boolean hasQuestion(Question question) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void deleteQuestion(Question target) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setQuestion(Question target, Question editedQuestion) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public ObservableList<Question> getFilteredQuestionList() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public ObservableList<Question> getFilteredQuizQuestionList() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void updateFilteredQuestionList(Predicate<Question> predicate) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void resetFilteredQuestionList() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void sortFilteredQuizQuestionList(Comparator<Question> comparator) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setTheme(Theme theme) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public Theme getTheme() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public ObservableList<TagStatistic> getTagStatistic() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void updateFilteredTagStatistic(Predicate<TagStatistic> predicate) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setPanel(String panel) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public String getPanel() {
-            return StatisticListPanel.STATISTIC_PANEL;
-        }
-
-        @Override
-        public void addNote(Note note) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void deleteNote(Note target) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setNote(Note target, Note editedNote) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public ObservableList<Note> getFilteredNoteList() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void updateFilteredNoteList(Predicate<Note> predicate) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void updateFilteredQuizQuestionList(Predicate<Question> predicate) {
-            throw new AssertionError("This method should not be called.");
-        }
     }
 
 }
